Server/requirements.txt
pycryptodome
uvicorn
fastapi[all]
python-multipart